% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2xy.R
\name{col2xy}
\alias{col2xy}
\title{Color to XY Conversion}
\usage{
col2xy(color)
}
\arguments{
\item{color}{color code}
}
\description{
Color to XY Conversion
}
\examples{
col2xy("#FF003F")
col2xy(rgb(221, 235, 30, maxColorValue = 255))
}
