% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_block.R
\name{col_block}
\alias{col_block}
\title{Display Color Code Blocks}
\usage{
col_block(color, bg = NULL)
}
\arguments{
\item{color}{color code}

\item{bg}{background color}
}
\description{
Display Color Code Blocks
}
\examples{
\dontrun{
col_block("#00af00")
col_block("#00af00", bg = "white")
col_block("#00af00", bg = "black")
}
}
